refactor(UserLoaderScreen): Move into withLoggedInUserSupport. Make class wait for token before fetc
