+++ lib/actions/user.js
@@ -20,59 +20,53 @@ function getStateForNewUser (auth0User) {
 }
 
 /**
- * Fetches user preferences or set initial values under state.user if no user has been loaded.
+ * Fetches user preferences to state.user, or set initial values under state.user if no user has been loaded.
  */
-export function ensureLoggedInUserIsFetched (auth) {
+export function fetchOrInitializeUser (auth) {
   return async function (dispatch, getState) {
-    const { otp, user } = getState()
-    const { loggedInUser } = user
+    const { otp } = getState()
+    const { accessToken, user } = auth
 
-    if (auth && !loggedInUser) {
-      const { accessToken, user } = auth
-      if (accessToken) {
-        try {
-          const result = await fetchUser(
-            otp.config.persistence.otp_middleware,
-            accessToken,
-            user.sub
-          )
+    try {
+      const result = await fetchUser(
+        otp.config.persistence.otp_middleware,
+        accessToken
+      )
 
-          // Beware! On AWS API gateway, if a user is not found in the middleware
-          // (e.g. they just created their Auth0 password but have not completed the account setup form yet),
-          // the call above will return, for example:
-          // {
-          //    status: 'success',
-          //    data: {
-          //      "result": "ERR",
-          //      "message": "No user with id=000000 found.",
-          //      "code": 404,
-          //      "detail": null
-          //    }
-          // }
-          //
-          // When not behind an API gateway (request sent directly to a middleware server),
-          // the same call will return:
-          // {
-          //    status: 'error',
-          //    message: 'Error get-ing user...'
-          // }
-          // TODO: Improve AWS response.
+      // Beware! On AWS API gateway, if a user is not found in the middleware
+      // (e.g. they just created their Auth0 password but have not completed the account setup form yet),
+      // the call above will return, for example:
+      // {
+      //    status: 'success',
+      //    data: {
+      //      "result": "ERR",
+      //      "message": "No user with id=000000 found.",
+      //      "code": 404,
+      //      "detail": null
+      //    }
+      // }
+      //
+      // The same call to a middleware instance that is not behind an API gateway
+      // will return:
+      // {
+      //    status: 'error',
+      //    message: 'Error get-ing user...'
+      // }
+      // TODO: Improve AWS response.
 
-          const resultData = result.data
-          const isNewAccount = result.status === 'error' || (resultData && resultData.result === 'ERR')
+      const resultData = result.data
+      const isNewAccount = result.status === 'error' || (resultData && resultData.result === 'ERR')
 
-          if (!isNewAccount) {
-            // TODO: Move next line somewhere else.
-            if (resultData.savedLocations === null) resultData.savedLocations = []
-            dispatch(setCurrentUser({ accessToken, user: resultData }))
-          } else {
-            dispatch(setCurrentUser({ accessToken, user: getStateForNewUser(auth.user) }))
-          }
-        } catch (error) {
-          // TODO: improve error handling.
-          alert(`An error was encountered:\n${error}`)
-        }
+      if (!isNewAccount) {
+        // TODO: Move next line somewhere else.
+        if (resultData.savedLocations === null) resultData.savedLocations = []
+        dispatch(setCurrentUser({ accessToken, user: resultData }))
+      } else {
+        dispatch(setCurrentUser({ accessToken, user: getStateForNewUser(user) }))
       }
+    } catch (error) {
+      // TODO: improve error handling.
+      alert(`An error was encountered:\n${error}`)
     }
   }
 }
